Remove assets related to Sirenix Odin Inspector plugin
Several plugin configuration files, asset files, and a readme related to the Sirenix Odin Inspector plugin were removed. This action indicates that the plugin is no longer being used in the project. The removal of these files helps to declutter the project hierarchy and cleans up any dependencies on this plugin.