Add crafting and consumable support to the project
This commit adds support for crafting and consumables to the project. It includes new prefabs, scripts and updates for crafting controller, consumable dictionary, and items. An instance has been introduced for CraftingController along with a ConsumableMatrix for crafting recipes. Sprite support has been added to the scriptable object definition of consumables. Furthermore, item controller and crafting prefab have been updated for better functionality.