Remove RecipeList.asset and Tooth_N_Nail.inputactions files
These files were deleted based on project needs. The RecipeList.asset file, part of weapon prefabs, and the Tooth_N_Nail.inputaction, which held settings for user input interactions, are no longer necessary in the codebase.

Also complete all the way up to the actual consumption of items to craft a new item in the crafting system.